$&dictforall works fully as of right now. there was an issue where pointers weren't getting appropriately forwarded, so there's functionally a use-after-free like scenario. solved by saving the offending (and required) list to a var

+++ gc.c
@@ -13,7 +13,7 @@ struct Space {
 	Space *next;
 };
 
-#define	SPACESIZE(sp)	(((sp)->top - (sp)->bot))
+#define	SPACESIZE(sp)	((size_t)((sp)->top - (sp)->bot))
 #define	SPACEFREE(sp)	(((sp)->top - (sp)->current))
 #define	SPACEUSED(sp)	(((sp)->current - (sp)->bot))
 #define	INSPACE(p, sp)	((sp)->bot <= (char *) (p) && (char *) (p) < (sp)->top)
